Added: setting and getting company wage item methods.
Added calls to addNewCompanyWageItem() and getCompanyWageItem(). The
<select> control does not work, it is for reference only. You will
still need to type in the wage item type into the text field. If you
look at the company model, it now contains a companyWageItem section
for HOURLY_PAY. But when I added the variables, the
addNewCompanyWageItem() blanks out the whole page. Still trying to
debug that. Hardcoding the values in the "companyWageItemModel" works.